requirements.txt
replicate
customtkinter
tkinter
pillow
urllib.request